terug naar 8 bits + comments eraan toegevoegd

+++ Gekke Arduino Game/src/Infrarood.h
@@ -6,11 +6,11 @@
 // Functies voor initialisatie en verwerking
 void initInfrarood();
 void processReceivedPulses();
-void sendDataByte(uint32_t data);  // Voeg deze functie toe om de data te verzenden
+void sendDataByte(uint8_t data);  // Voeg deze functie toe om de data te verzenden
 
 // Externe variabelen
-extern uint32_t tempByte;
+extern uint8_t tempByte;
 extern bool newDataAvailable; // Vlag voor nieuwe data
-extern uint16_t readCount;
+extern uint8_t readCount;
 
 #endif
